backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlmodel==0.0.22
SQLAlchemy==2.0.36
psycopg2-binary

# Env & HTTP
python-dotenv==1.0.1
httpx==0.24.1

# OpenAI Agents SDK (Phase 5 core)
openai-agents
